Remove the initial card model that served as a prototype and replace it with the more appropriate, commander model
